requirements.txt
astroid==2.3.2
autopep8==1.4.4
coverage==4.5.4
entrypoints==0.3
flake8==3.7.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
nltk==3.4.5
pdf2image==1.9.0
Pillow==6.2.1
pudb==2019.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.3
pytesseract==0.3.0
six==1.12.0
urwid==2.0.1
wrapt==1.11.2
